Massive hatch.config rewrite
Moved everything away from the lazy-config stuff that was previously
used. Hatch was already parsing the whole toml file anyway, it didn't
make much sense to do lazy loading of the configuration.

I've tried to keep the interface that was used almost the same, though
it is a bit different internally. Instead of passing a `dict` the
classes take a set of keyword arguments. All of these are type-checked
with [pydantic][1]. This cleans the code in hatch.config.model up a lot,
removing tons of duplicated type checks for the different parts of the
configuration. Now it's all basically taken care of, including the use
of `typing.Literal`s for things like `mode` and `style`.

Incidentally, this also fixes the issue of the spinner being
un-configurable, there's part of this that adds it as a real config so
users can have the cool `dots8Bit` spinner.

[1]: https://github.com/pydantic/pydantic