Fix spinner style settings
Allows for customizing the spinner style finally! Previously it had a
problem where the attribute name was not `_style_level_spinner`, so it
was never getting set from the `initialize_styles` method. This made the
setting in the user's `config.toml` never see the light of day.

To fix this I also changed all of the `style_level_{level}` names to be
just `style_{level}`, cleaning up a bit of the code.

And for some reason the `Spinner` class doesn't keep a `name` attribute,
so I'll be poking the rich project about that in a minute, since I
basically had to create a `Spinner()` class with no ability to use it.