Env version-source: add a fallback value
Our build system sets the package version number via an environment variable.

When developers are building locally (eg when doing a `uv sync`) the
environment variable will not be set and the build fails. This is inconvenient
and we'd rather it just fell back to some default value like 0.0.1.dev0+local)