fixing multiple patch for a module
As mentioned in the issue #1871 if happen that a module is distributed across multiple location all location should be included in .pth file. In this commit is the fix for that. Current behavior for this scenario is to overwrite and keep only the last one.
P.S: ruff to an auto-format over this file. Hope doesn't hurt or I simply revert ruff autoformatting.